move gossipsub into a separate crate
