+++ beacon_node/lighthouse_network/Cargo.toml
@@ -5,8 +5,8 @@ authors = ["Sigma Prime <contact@sigmaprime.io>"]
 edition = { workspace = true }
 
 [dependencies]
-async-channel = { workspace = true }
 discv5 = { workspace = true }
+gossipsub = { workspace = true }
 unsigned-varint = { version = "0.6", features = ["codec"] }
 ssz_types = { workspace = true }
 types = { workspace = true }
@@ -50,16 +50,12 @@ either = { workspace = true }
 
 # Local dependencies
 futures-ticker = "0.0.3"
-futures-timer = "3.0.2"
 getrandom = "0.2.11"
 hex_fmt = "0.3.0"
 instant = "0.1.12"
-quick-protobuf = "0.8"
 void = "1.0.2"
-asynchronous-codec = "0.7.0"
 base64 = "0.21.5"
 libp2p-mplex = "0.41"
-quick-protobuf-codec = "0.3"
 
 [dependencies.libp2p]
 version = "0.53"
@@ -72,7 +68,7 @@ slog-async = { workspace = true }
 tempfile = { workspace = true }
 quickcheck = { workspace = true }
 quickcheck_macros = { workspace = true }
-async-std = { version = "1.6.3", features = ["unstable"] }
+async-channel = { workspace = true }
 
 [features]
 libp2p-websocket = []
