+++ beacon_node/lighthouse_network/gossipsub/src/behaviour/tests.rs
@@ -21,17 +21,17 @@
 // Collection of tests for the gossipsub network behaviour
 
 use super::*;
-use crate::gossipsub::subscription_filter::WhitelistSubscriptionFilter;
-use crate::gossipsub::transform::{DataTransform, IdentityTransform};
-use crate::gossipsub::types::{RpcOut, RpcReceiver};
-use crate::gossipsub::ValidationError;
-use crate::gossipsub::{
+use crate::subscription_filter::WhitelistSubscriptionFilter;
+use crate::transform::{DataTransform, IdentityTransform};
+use crate::types::{RpcOut, RpcReceiver};
+use crate::ValidationError;
+use crate::{
     config::Config, config::ConfigBuilder, types::Rpc, IdentTopic as Topic, TopicScoreParams,
 };
-use async_std::net::Ipv4Addr;
 use byteorder::{BigEndian, ByteOrder};
 use libp2p::core::{ConnectedPoint, Endpoint};
 use rand::Rng;
+use std::net::Ipv4Addr;
 use std::thread::sleep;
 use std::time::Duration;
