+++ beacon_node/lighthouse_network/src/lib.rs
@@ -10,7 +10,6 @@ pub mod service;
 
 #[allow(clippy::mutable_key_type)] // PeerId in hashmaps are no longer permitted by clippy
 pub mod discovery;
-pub mod gossipsub;
 pub mod listen_addr;
 pub mod metrics;
 pub mod peer_manager;
