+++ beacon_node/lighthouse_network/src/service/gossipsub_scoring_parameters.rs
@@ -1,9 +1,9 @@
-use crate::gossipsub::{
+use crate::types::{GossipEncoding, GossipKind, GossipTopic};
+use crate::{error, TopicHash};
+use gossipsub::{
     Config as GossipsubConfig, IdentTopic as Topic, PeerScoreParams, PeerScoreThresholds,
     TopicScoreParams,
 };
-use crate::types::{GossipEncoding, GossipKind, GossipTopic};
-use crate::{error, TopicHash};
 use std::cmp::max;
 use std::collections::HashMap;
 use std::marker::PhantomData;
