+++ beacon_node/lighthouse_network/src/service/utils.rs
@@ -1,4 +1,3 @@
-use crate::gossipsub;
 use crate::multiaddr::Protocol;
 use crate::rpc::{MetaData, MetaDataV1, MetaDataV2};
 use crate::types::{
@@ -6,6 +5,7 @@ use crate::types::{
 };
 use crate::{GossipTopic, NetworkConfig};
 use futures::future::Either;
+use gossipsub;
 use libp2p::core::{multiaddr::Multiaddr, muxing::StreamMuxerBox, transport::Boxed};
 use libp2p::identity::{secp256k1, Keypair};
 use libp2p::{core, noise, yamux, PeerId, Transport};
